Crate yaserde

YaSerDe

YaSerDe is a framework for serializing and deserializing Rust data structures efficiently and generically from and into XML.

Modules

de

Generic data structure deserialization framework.

ser

Generic data structure serialization framework.

Macros

deserialize_and_validate

serialize_and_validate

test_for_attribute_type

test_for_type

Traits

Visitor

A visitor that can be implemented to retrieve information from source file.

YaDeserialize

A data structure that can be deserialized from any data format supported by YaSerDe.

YaSerialize

A data structure that can be serialized into any data format supported by YaSerDe.